project/requirements.txt
Django==5.0.7
celery==5.4.0
redis==5.0.7
asgiref==3.8.1
django-debug-toolbar==5.1.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
djangorestframework==3.16.0

black==24.10.0
ruff==0.6.9
mypy==1.5.0
drf-yasg==1.21.10
parameterized==0.9.0
flower==1.2.0

